Modify trie.c for partial but not unique match search
Added a new function to trie.c. TrieSearchPartial() was renamed to TrieSearchPartialUnique() to preserve the partial but unique match search functionality. TrieSearchPartial() was tweaked to allow non-unique partial matches to be returned. This makes places_server more useful because it will find matches to a users search query more frequently.