Clean up repo and places_svc.c and places_server.c
Cleaned up and commented code in places_svc.c and places_server.c. Also removed all of the object, executable, and temp files (the ones ending in '~') from git tracking. The build files should not be tracked but rather built fresh after a pull. The temp files seemed to be some leftover garbage from a "git add ." They were not used and cluttered up the repo.