Modify places_server to pass back found city along with rest of airports
places_server now passes back the found city/state and lat long information from the trie search. places_client was changed to print this out before printing the airport data. This change required an update to the trie structure to store the original entry name as a value in each terminating node.